fix: PiPが暗転から始まるのを修正。PiP->Display2に切り替わらない不具合を修正。refactor: PiPHandlerを別ファイルに分離
